Whatever floats your boat * Transitioned to floating point angles * Complete rewrite of the variable system, now no variables are actually sent ( and loaded ) over the wire. Aside from being bad practice ( and a potential safety hazard ), this is now done by instructing the interpreter how to build the variable array, rather than actually sending said variables. * Implemented a target system * A lot of the code is kept for legacy, a cleanup is due soon. * Fixed instructions with multiple spaces between parameters being iffy * Removed header containing metadata

+++ Assembler/inc/Instruction.hpp
@@ -4,5 +4,5 @@
 
 struct Instruction{
     uint8_t opcode;
-    int32_t params[10];
+    float params[10];
 };
